replaced temperature type from float to double

+++ src/main/java/ru/hehmdalolkek/weather/dto/TemperatureResponse.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @Builder
 public class TemperatureResponse {
 
-    private float temperature;
+    private double temperature;
 
     private LocalDateTime datetime;
